keep-class-files option added

+++ src/Commands/ConvertToMultiFileComponent.php
@@ -13,9 +13,11 @@ use LivewireV4\Utility\ViewFilePath;
 
 class ConvertToMultiFileComponent extends Command
 {
-    protected $signature = 'convert-class-to:mfc {path}';
-
-    protected $description = 'Convert Livewire class based component to Multi File Component(MFC)';
+    protected $signature = 'convert-class-to:mfc 
+                            {path : Path to the component or directory}
+                            {--keep-class-files : Keep original class files after conversion}';
+    
+    protected $description = 'Convert Livewire class components to MFC';
 
     public function handle(): void
     {
@@ -88,18 +90,22 @@ class ConvertToMultiFileComponent extends Command
 
         $this->info('DONE !!!');
 
-        $viewFilePath = ViewFilePath::make()->path($fullPath)->viewFilePath();
-        File::delete($fullPath);
-        File::delete($viewFilePath);
+        //if user wants to keep the old class files
+        if(!$this->option('keep-class-files')){
+            $viewFilePath = ViewFilePath::make()->path($fullPath)->viewFilePath();
+            File::delete($fullPath);
+            File::delete($viewFilePath);
 
-        // check if the base directory is empty
-        if (File::isEmptyDirectory(dirname($fullPath))) {
-            File::deleteDirectory(dirname($fullPath));
-            $this->info('Cleaning the Empty Componet Directoy');
-        }
-        if (File::isEmptyDirectory(dirname($viewFilePath))) {
-            File::deleteDirectory(dirname($viewFilePath));
-            $this->info('Cleaning the Empty View Directoy');
+            // check if the base directory is empty
+            if (File::isEmptyDirectory(dirname($fullPath))) {
+                File::deleteDirectory(dirname($fullPath));
+                $this->info('Cleaning the Empty Componet Directoy');
+            }
+            if (File::isEmptyDirectory(dirname($viewFilePath))) {
+                File::deleteDirectory(dirname($viewFilePath));
+                $this->info('Cleaning the Empty View Directoy');
+            }
+
         }
 
         $this->newLine();
